cleanup test_orbital_core unit testing module
